+++ src/renderer/hooks/useRequestUserProfile.js
@@ -7,9 +7,9 @@ import {
   requestGitlabUserProfile,
 } from '../service/api';
 
-import fetchReducer from '../fetchReducer';
+import fetchReducer from '../reducers/fetchReducer';
 
-export function useRequestUserProfile(selectedProvider, token) {
+export default function useRequestUserProfile(selectedProvider, token) {
   const [state, dispatch] = useReducer(fetchReducer, {
     isLoading: false,
     isError: false,
